fix: grouped domain objects into common modules
