--- proto/src/domain/block_inputs.rs
@@ -1,66 +1,0 @@
-use miden_crypto::merkle::MmrPeaks;
-use miden_objects::BlockHeader;
-
-use crate::{
-    domain::{
-        account_input_record::AccountInputRecord, nullifier_input_record::NullifierInputRecord,
-        try_convert,
-    },
-    errors::ParseError,
-    responses,
-};
-
-/// Information needed from the store to build a block
-#[derive(Clone, Debug)]
-pub struct BlockInputs {
-    /// Previous block header
-    pub block_header: BlockHeader,
-
-    /// MMR peaks for the current chain state
-    pub chain_peaks: MmrPeaks,
-
-    /// The hashes of the requested accounts and their authentication paths
-    pub account_states: Vec<AccountInputRecord>,
-
-    /// The requested nullifiers and their authentication paths
-    pub nullifiers: Vec<NullifierInputRecord>,
-}
-
-// FROM
-// ================================================================================================
-
-impl TryFrom<responses::GetBlockInputsResponse> for BlockInputs {
-    type Error = ParseError;
-
-    fn try_from(get_block_inputs: responses::GetBlockInputsResponse) -> Result<Self, Self::Error> {
-        let block_header: BlockHeader = get_block_inputs
-            .block_header
-            .ok_or(ParseError::ProtobufMissingData)?
-            .try_into()?;
-
-        let chain_peaks = {
-            // setting the number of leaves to the current block number gives us one leaf less than
-            // what is currently in the chain MMR (i.e., chain MMR with block_num = 1 has 2 leave);
-            // this is because GetBlockInputs returns the state of the chain MMR as of one block
-            // ago so that block_header.chain_root matches the hash of MMR peaks.
-            let num_leaves = block_header.block_num() as usize;
-
-            MmrPeaks::new(
-                num_leaves,
-                get_block_inputs
-                    .mmr_peaks
-                    .into_iter()
-                    .map(|peak| peak.try_into())
-                    .collect::<Result<_, Self::Error>>()?,
-            )
-            .map_err(Self::Error::MmrPeaksError)?
-        };
-
-        Ok(Self {
-            block_header,
-            chain_peaks,
-            account_states: try_convert(get_block_inputs.account_states)?,
-            nullifiers: try_convert(get_block_inputs.nullifiers)?,
-        })
-    }
-}
