+++ proto/src/domain/digest.rs
@@ -1,3 +1,6 @@
+use std::fmt::{Debug, Display, Formatter};
+
+use hex::{FromHex, ToHex};
 use miden_crypto::{Felt, StarkField};
 use miden_objects::{
     notes::{NoteId, Nullifier},
@@ -5,6 +8,89 @@ use miden_objects::{
 };
 
 use crate::{digest, errors::ParseError};
+
+// CONSTANTS
+// ================================================================================================
+
+pub const DIGEST_DATA_SIZE: usize = 32;
+
+// FORMATTING
+// ================================================================================================
+
+impl Display for digest::Digest {
+    fn fmt(
+        &self,
+        f: &mut Formatter<'_>,
+    ) -> std::fmt::Result {
+        f.write_str(&self.encode_hex::<String>())
+    }
+}
+
+impl Debug for digest::Digest {
+    fn fmt(
+        &self,
+        f: &mut Formatter<'_>,
+    ) -> std::fmt::Result {
+        Display::fmt(self, f)
+    }
+}
+
+impl ToHex for &digest::Digest {
+    fn encode_hex<T: FromIterator<char>>(&self) -> T {
+        (*self).encode_hex()
+    }
+
+    fn encode_hex_upper<T: FromIterator<char>>(&self) -> T {
+        (*self).encode_hex_upper()
+    }
+}
+
+impl ToHex for digest::Digest {
+    fn encode_hex<T: FromIterator<char>>(&self) -> T {
+        let mut data: Vec<char> = Vec::with_capacity(DIGEST_DATA_SIZE);
+        data.extend(format!("{:016x}", self.d0).chars());
+        data.extend(format!("{:016x}", self.d1).chars());
+        data.extend(format!("{:016x}", self.d2).chars());
+        data.extend(format!("{:016x}", self.d3).chars());
+        data.into_iter().collect()
+    }
+
+    fn encode_hex_upper<T: FromIterator<char>>(&self) -> T {
+        let mut data: Vec<char> = Vec::with_capacity(DIGEST_DATA_SIZE);
+        data.extend(format!("{:016X}", self.d0).chars());
+        data.extend(format!("{:016X}", self.d1).chars());
+        data.extend(format!("{:016X}", self.d2).chars());
+        data.extend(format!("{:016X}", self.d3).chars());
+        data.into_iter().collect()
+    }
+}
+
+impl FromHex for digest::Digest {
+    type Error = ParseError;
+
+    fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
+        let data = hex::decode(hex)?;
+
+        match data.len() {
+            size if size < DIGEST_DATA_SIZE => Err(ParseError::InsufficientData {
+                expected: DIGEST_DATA_SIZE,
+                got: size,
+            }),
+            size if size > DIGEST_DATA_SIZE => Err(ParseError::TooMuchData {
+                expected: DIGEST_DATA_SIZE,
+                got: size,
+            }),
+            _ => {
+                let d0 = u64::from_be_bytes(data[..8].try_into().unwrap());
+                let d1 = u64::from_be_bytes(data[8..16].try_into().unwrap());
+                let d2 = u64::from_be_bytes(data[16..24].try_into().unwrap());
+                let d3 = u64::from_be_bytes(data[24..32].try_into().unwrap());
+
+                Ok(digest::Digest { d0, d1, d2, d3 })
+            },
+        }
+    }
+}
 
 // INTO
 // ================================================================================================
@@ -134,5 +220,51 @@ impl TryFrom<&digest::Digest> for Digest {
 
     fn try_from(value: &digest::Digest) -> Result<Self, Self::Error> {
         value.clone().try_into()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use hex::{self, FromHex, ToHex};
+    use proptest::prelude::*;
+
+    use crate::digest::Digest;
+
+    #[test]
+    fn test_hex_digest() {
+        let digest = Digest {
+            d0: 3488802789098113751,
+            d1: 5271242459988994564,
+            d2: 17816570245237064784,
+            d3: 10910963388447438895,
+        };
+        let encoded: String = ToHex::encode_hex(&digest);
+        let round_trip: Result<Digest, _> = FromHex::from_hex::<&[u8]>(encoded.as_ref());
+        assert_eq!(digest, round_trip.unwrap());
+
+        let digest = Digest {
+            d0: 0,
+            d1: 0,
+            d2: 0,
+            d3: 0,
+        };
+        let encoded: String = ToHex::encode_hex(&digest);
+        let round_trip: Result<Digest, _> = FromHex::from_hex::<&[u8]>(encoded.as_ref());
+        assert_eq!(digest, round_trip.unwrap());
+    }
+
+    proptest! {
+        #[test]
+        fn test_encode_decode(
+            d0: u64,
+            d1: u64,
+            d2: u64,
+            d3: u64,
+        ) {
+            let digest = Digest { d0, d1, d2, d3 };
+            let encoded: String = ToHex::encode_hex(&digest);
+            let round_trip: Result<Digest, _> = FromHex::from_hex::<&[u8]>(encoded.as_ref());
+            assert_eq!(digest, round_trip.unwrap());
+        }
     }
 }
