+++ proto/src/domain/notes.rs
@@ -2,7 +2,22 @@ use miden_objects::notes::NoteEnvelope;
 
 use crate::note;
 
-// INTO
+// Note
+// ================================================================================================
+
+impl From<note::Note> for note::NoteSyncRecord {
+    fn from(value: note::Note) -> Self {
+        Self {
+            note_index: value.note_index,
+            note_hash: value.note_hash,
+            sender: value.sender,
+            tag: value.tag,
+            merkle_path: value.merkle_path,
+        }
+    }
+}
+
+// NoteCreated
 // ================================================================================================
 
 impl From<(u64, NoteEnvelope)> for note::NoteCreated {
